classroom-attendance/backend/requirements.txt
Flask
firebase-admin
deepface
opencv-python
numpy
pillow
